enhance(runtime): prepare for the future supergraph executor
