+++ .changeset/light-badgers-tan.md
@@ -1,0 +1,5 @@
+---
+'@graphql-mesh/fusion-runtime': patch
+---
+
+Refactor to make it easier to replace the supergraph execution
